requirement.txt
python-dotenv==1.2.2
psycopg2==2.9.11
clickhouse-connect==0.11.0
kafka-python==2.3.0